+++ NEWS.md
@@ -61,6 +61,7 @@
 * `with_defaults()` handles automatic naming of very long arguments correctly (#774, @michaelchirico)
 * `object_name_linter()` no longer lints names used for subsetting (#582, @AshesITR)
 * `lint_package()` warns and returns `NULL` if no package is found (instead of giving a peculiar error message) (#776, @michaelchirico)
+* `lint_package()` is also stricter about what it considers to be a package -- folders named `DESCRIPTION` are ignored (#702, @michaelchirico)
 
 # lintr 2.0.1